Try to gracefully terminate the application when the windowtitle is known
This change only affects the Windows build.
Before, the watchdog always hard terminated the watched application when
a stop or restart command was received.
Now and only when the window title is given as an argument, the watchdog
tries to gracefully close the watched application by sending a WM_Close
event.
If the ‚windowtitle‘ argument is not set it uses the old behaviour and
terminates the application.
Also added missing ‚windowtitle‘ argument documentation and describe the
new behaviour.