improve handling of close application with window title given
This changes the behaviour when the window title is not found.
Old behaviour was if the window title was not found it just returned
and set the application state falsy to terminated.
Now it always terminates the watched application and updates the
application status.
Also it fixes a possible handle leak when using WM_Sent to close the
watched application.